# openapi_v3.json
{
  "components": {
    "schemas": {
...
      "Error": {
        "properties": {
          "error": {
            "description": "Error type",
            "example": "Not Found",
            "type": "string"
          },
          "message": {
            "description": "Detailed error message",
            "example": "The requested resource was not found",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "World": {
        "properties": {
          "creationdate": {
            "description": "Date when the world was created",
            "example": "2023-01-01",
            "format": "date",
            "type": "string"
          },
          "id": {
            "description": "Unique identifier for the world",
            "example": 1,
            "type": "integer"
          },
          "name": {
            "description": "Name of the world",
            "example": "World 1-1",
            "type": "string"
          }
        },
        "type": "object"
      },
      "WorldInput": {
        "properties": {
          "name": {
            "description": "Name of the new world",
            "example": "World 1-1",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "type": "object"
      },
      "WorldResponse": {
        "properties": {
          "world": {
            "description": "Array of world objects",
            "items": {
              "$ref": "#/components/schemas/World"
            },
            "type": "array"
          }
        },
        "type": "object"
      }
    }
  },
  "info": {
    "contact": {
      "name": "QARATE API Support",
      "url": "https://youtu.be/q9Xoic_14M0"
    },
    "description": "Enhanced API with automatic contract generation\n\u0414\u0435\u043c\u043e API \u0434\u043b\u044f \u0434\u0435\u043c\u043e\u043d\u0441\u0442\u0440\u0430\u0446\u0438\u0438 \u0432\u043e\u0437\u043c\u043e\u0436\u043d\u043e\u0441\u0442\u0435\u0439 Postman.\n\u0421\u043e\u0437\u0434\u0430\u043d\u043e \u0441\u043f\u0435\u0446\u0438\u0430\u043b\u044c\u043d\u043e \u0434\u043b\u044f \u0432\u0435\u0431\u0438\u043d\u0430\u0440\u043e\u0432\nQARATE #5: https://youtu.be/q9Xoic_14M0\nQARATE #6: https://youtu.be/WVNVeHtmBjc",
    "title": "BowserAPI",
    "version": "1.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
    "/addworld": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorldInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "examples": {
                  "world_created": {
                    "value": {
                      "world": [
                        {
                          "creationdate": "2023-01-01",
                          "id": 2,
                          "name": "World 1-2"
                        }
                      ]
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/WorldResponse"
                }
              }
            },
            "description": "\u041c\u0438\u0440 \u0443\u0441\u043f\u0435\u0448\u043d\u043e \u0434\u043e\u0431\u0430\u0432\u043b\u0435\u043d"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "\u041d\u0435\u0432\u0435\u0440\u043d\u044b\u0435 \u0434\u0430\u043d\u043d\u044b\u0435 \u0437\u0430\u043f\u0440\u043e\u0441\u0430"
          }
        },
        "summary": "\u0414\u043e\u0431\u0430\u0432\u043b\u044f\u0435\u0442 \u043d\u043e\u0432\u044b\u0439 \u043c\u0438\u0440",
        "tags": [
          "world"
        ]
      }
    },
...
    "/world": {
      "get": {
        "parameters": [
          {
            "description": "ID \u043a\u043e\u043d\u043a\u0440\u0435\u0442\u043d\u043e\u0433\u043e \u043c\u0438\u0440\u0430 (\u043e\u043f\u0446\u0438\u043e\u043d\u0430\u043b\u044c\u043d\u043e)",
            "in": "path",
            "name": "worldid",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "examples": {
                  "world_example": {
                    "value": {
                      "world": [
                        {
                          "creationdate": "2023-01-01",
                          "id": 1,
                          "name": "World 1-1"
                        }
                      ]
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/WorldResponse"
                }
              }
            },
            "description": "\u041c\u0438\u0440(\u044b) \u0443\u0441\u043f\u0435\u0448\u043d\u043e \u0432\u043e\u0437\u0432\u0440\u0430\u0449\u0435\u043d(\u044b)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "\u041c\u0438\u0440 \u043d\u0435 \u043d\u0430\u0439\u0434\u0435\u043d"
          }
        },
        "summary": "\u0412\u043e\u0437\u0432\u0440\u0430\u0449\u0430\u0435\u0442 \u0438\u043d\u0444\u043e\u0440\u043c\u0430\u0446\u0438\u044e \u043e \u043c\u0438\u0440\u0435/\u043c\u0438\u0440\u0430\u0445",
        "tags": [
          "world"
        ]
      }
    },
    "/world/{worldid}": {
      "delete": {
        "parameters": [
          {
            "description": "ID \u043c\u0438\u0440\u0430 \u0434\u043b\u044f \u0443\u0434\u0430\u043b\u0435\u043d\u0438\u044f",
            "in": "path",
            "name": "worldid",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "\u041c\u0438\u0440 \u0443\u0441\u043f\u0435\u0448\u043d\u043e \u0443\u0434\u0430\u043b\u0435\u043d"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "\u041c\u0438\u0440 \u043d\u0435 \u043d\u0430\u0439\u0434\u0435\u043d"
          }
        },
        "summary": "\u0423\u0434\u0430\u043b\u044f\u0435\u0442 \u0432\u044b\u0431\u0440\u0430\u043d\u043d\u044b\u0439 \u043c\u0438\u0440",
        "tags": [
          "world"
        ]
      },
      "get": {
        "parameters": [
          {
            "description": "ID \u043a\u043e\u043d\u043a\u0440\u0435\u0442\u043d\u043e\u0433\u043e \u043c\u0438\u0440\u0430 (\u043e\u043f\u0446\u0438\u043e\u043d\u0430\u043b\u044c\u043d\u043e)",
            "in": "path",
            "name": "worldid",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "examples": {
                  "world_example": {
                    "value": {
                      "world": [
                        {
                          "creationdate": "2023-01-01",
                          "id": 1,
                          "name": "World 1-1"
                        }
                      ]
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/WorldResponse"
                }
              }
            },
            "description": "\u041c\u0438\u0440(\u044b) \u0443\u0441\u043f\u0435\u0448\u043d\u043e \u0432\u043e\u0437\u0432\u0440\u0430\u0449\u0435\u043d(\u044b)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "\u041c\u0438\u0440 \u043d\u0435 \u043d\u0430\u0439\u0434\u0435\u043d"
          }
        },
        "summary": "\u0412\u043e\u0437\u0432\u0440\u0430\u0449\u0430\u0435\u0442 \u0438\u043d\u0444\u043e\u0440\u043c\u0430\u0446\u0438\u044e \u043e \u043c\u0438\u0440\u0435/\u043c\u0438\u0440\u0430\u0445",
        "tags": [
          "world"
        ]
      },
      "put": {
        "parameters": [
          {
            "description": "ID \u043c\u0438\u0440\u0430 \u0434\u043b\u044f \u043e\u0431\u043d\u043e\u0432\u043b\u0435\u043d\u0438\u044f",
            "in": "path",
            "name": "worldid",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorldInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "examples": {
                  "world_updated": {
                    "value": {
                      "world": [
                        {
                          "creationdate": "2023-01-01",
                          "id": 1,
                          "name": "Updated World Name"
                        }
                      ]
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/WorldResponse"
                }
              }
            },
            "description": "\u041c\u0438\u0440 \u0443\u0441\u043f\u0435\u0448\u043d\u043e \u043e\u0431\u043d\u043e\u0432\u043b\u0435\u043d"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "\u041c\u0438\u0440 \u043d\u0435 \u043d\u0430\u0439\u0434\u0435\u043d"
          }
        },
        "summary": "\u041e\u0431\u043d\u043e\u0432\u043b\u044f\u0435\u0442 \u0432\u044b\u0431\u0440\u0430\u043d\u043d\u044b\u0439 \u043c\u0438\u0440",
        "tags": [
          "world"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Local development server",
      "url": "/"
    }
  ]
}